Clean up CFI sanitizer flags in static libflac.
Remove diagnostics mode from CFI across the board. This should reduce
performance overhead and also allows the minimal runtime to work when
other ubsan sanitizers are enabled. CFI stack dumps should include a
CFI related function, so it should be apparent when a crash is
CFI-related.

Since we no longer need the diagnostics flag, we can remove the CFI flag
as well since static libraries don't need this flag to be built with CFI
(a CFI version gets built if a dependency has it enabled).

Bug: 117417735
Test: make -j
Change-Id: I3e4a36b0f3f5fe953c352b736c04b67cdaaa7ae0